+++ lib/karafka/setup/config.rb
@@ -114,6 +114,13 @@ module Karafka
         #   instances
         setting :process, default: Process.new
 
+        # Namespace for CLI related settings
+        setting :cli do
+          # option contract [Object] cli setup validation contract (in the context of options and
+          # topics)
+          setting :contract, default: Contracts::ServerCliOptions.new
+        end
+
         setting :routing do
           # option builder [Karafka::Routing::Builder] builder instance
           setting :builder, default: Routing::Builder.new